V0.0.3 Finished
Market C/S part
messsagebox upgrade
collectMaterial add message in PathMessage
Others.....

+++ ResourceEmperorServer/REProtocol/ItemID.cs
@@ -5,7 +5,7 @@ using System.Text;
 
 namespace REProtocol
 {
-    public enum ItemID : ushort
+    public enum ItemID : int
     {
         MaterialBegin = 0,
         Log = 0,
